C++: reorganize core source files
* Merge NativeStackTrace.cpp into Support.cpp (less files is better)
* Rename _config.h -> Config.h (not an internal header)
* Rename uPlatform/WinAPIHelper.h -> uno/WinAPI.h (used by uno/ files)
* Rename uInitRtti -> uInitTypes (more specific name)
* Remove unused header files
