+++ lib/UnoCore/UnoCore.unoproj
@@ -98,25 +98,22 @@
     "cpp/main-mobile.cpp:file",
     "cpp/uno.cpp.uxl:extensions",
     "cpp/uno.h:file",
-    "cpp/uno/_config.h:file",
+    "cpp/uno/Config.h:file",
     "cpp/uno/_internal.h:file",
     "cpp/uno/_invoke.cpp:file",
-    "cpp/uno/graphics/GLHelper.h:file",
     "cpp/uno/Memory.cpp:file",
     "cpp/uno/Memory.h:file",
-    "cpp/uno/NativeStackTrace.cpp:file",
     "cpp/uno/ObjectModel.cpp:file",
     "cpp/uno/ObjectModel.h:file",
     "cpp/uno/ObjectMonitor.h:file",
     "cpp/uno/Reflection.cpp:file",
     "cpp/uno/Reflection.h:file",
     "cpp/uno/Support.cpp:file",
-    "cpp/uno/Support.h:file",
-    "cpp/uno/SupportApple.mm:file",
-    "cpp/uno/uno.h:file",
+    "cpp/uno/Support.mm:file",
+    "cpp/uno/ThreadUtils.h:file",
     "cpp/uPlatform/GLHelper.h:file",
     "cpp/uPlatform/GraphicsContext.h:file",
-    "cpp/uPlatform/WinAPIHelper.h:file",
+    "cpp/uno/WinAPI.h:file",
     "ios/@(Icons)/Contents.json:file",
     "ios/@(Project.Name).xcodeproj/project.pbxproj:file",
     "ios/@(Project.Name).xcodeproj/project.xcworkspace/contents.xcworkspacedata:file",
