+++ lib/UnoCore/cpp/main-mobile.cpp
@@ -8,9 +8,9 @@ int uArgc = 0;
 char** uArgv = nullptr;
 
 @(TypeObjects.Declaration:JoinSorted())
-void uInitRtti(uType*(*factories[])());
+void uInitTypes(uType*(*factories[])());
 
-void uInitRtti()
+void uInitTypes()
 {
     static uType*(*factories[])() =
     {
@@ -18,7 +18,7 @@ void uInitRtti()
         nullptr
     };
 
-    uInitRtti(factories);
+    uInitTypes(factories);
 }
 
 @(Main.Include:JoinSorted('\n', '#include <', '>'))
