+++ lib/UnoCore/cpp/uno.cpp.uxl
@@ -2,34 +2,30 @@
 
     <Require Template="XliPlatform" />
 
+    <!-- Core source code -->
     <ProcessFile HeaderFile="uno.h" />
-    <ProcessFile HeaderFile="uno/_config.h" />
     <ProcessFile HeaderFile="uno/_internal.h" />
     <ProcessFile SourceFile="uno/_invoke.cpp" />
+    <ProcessFile HeaderFile="uno/Config.h" />
     <ProcessFile SourceFile="uno/Memory.cpp" />
     <ProcessFile HeaderFile="uno/Memory.h" />
-    <ProcessFile SourceFile="uno/NativeStackTrace.cpp" />
     <ProcessFile SourceFile="uno/ObjectModel.cpp" />
     <ProcessFile HeaderFile="uno/ObjectModel.h" />
     <ProcessFile HeaderFile="uno/ObjectMonitor.h" />
-    <ProcessFile SourceFile="uno/Support.cpp" />
-    <ProcessFile HeaderFile="uno/Support.h" />
-    <ProcessFile SourceFile="uno/SupportApple.mm" Condition="APPLE" />
     <ProcessFile SourceFile="uno/Reflection.cpp" Condition="REFLECTION" />
     <ProcessFile HeaderFile="uno/Reflection.h" Condition="REFLECTION" />
+    <ProcessFile SourceFile="uno/Support.cpp" />
+    <ProcessFile SourceFile="uno/Support.mm" Condition="APPLE" />
+    <ProcessFile HeaderFile="uno/ThreadUtils.h" />
+    <ProcessFile HeaderFile="uno/WinAPI.h" Condition="WIN32" />
 
-    <!-- Platform headers -->
-    <ProcessFile HeaderFile="uPlatform/GraphicsContext.h" Condition="!MOBILE" />
+    <!-- Graphics headers -->
+    <ProcessFile HeaderFile="uPlatform/GraphicsContext.h" Condition="!MOBILE && OPENGL" />
     <ProcessFile HeaderFile="uPlatform/GLHelper.h" Condition="OPENGL" />
-    <ProcessFile HeaderFile="uPlatform/WinAPIHelper.h" Condition="WIN32" />
 
     <!-- Entrypoint -->
     <ProcessFile SourceFile="main-app.cpp" Condition="!MOBILE && !CONSOLE"/>
     <ProcessFile SourceFile="main-console.cpp" Condition="!MOBILE && CONSOLE"/>
     <ProcessFile SourceFile="main-mobile.cpp" Condition="MOBILE" />
-
-    <!-- Deprecated headers -->
-    <ProcessFile HeaderFile="uno/graphics/GLHelper.h" Condition="ANDROID" />
-    <ProcessFile HeaderFile="uno/uno.h" Condition="APPLE" />
 
 </Extensions>
