+++ lib/UnoCore/cpp/uno.h
@@ -2,4 +2,4 @@
 // @(MSG_EDIT_WARNING)
 
 #include <uno/ObjectModel.h>
-#include <uno/Support.h>
+#include <uno/ThreadUtils.h>
