+++ lib/UnoCore/cpp/uno/WinAPI.h
