--- lib/UnoCore/cpp/uno/graphics/GLHelper.h
@@ -1,5 +1,0 @@
-// @(MSG_ORIGIN)
-// @(MSG_EDIT_WARNING)
-
-#warning "Deprecated: Please include <uDroid/GLHelper.h> instead."
-#include <uDroid/GLHelper.h>
