--- lib/UnoCore/cpp/uno/uno.h
@@ -1,5 +1,0 @@
-// @(MSG_ORIGIN)
-// @(MSG_EDIT_WARNING)
-
-//#warning "Deprecated: Please include <uno.h> instead."
-#include "../uno.h"
