+++ lib/UnoCore/src/Uno/IO/FileStatusHelpers.Windows.uno
@@ -4,7 +4,7 @@ using Uno.Time;
 
 namespace Uno.IO
 {
-    [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
+    [Require("Source.Include", "uno/WinAPI.h")]
     extern(MSVC) internal static class FileStatusHelpers
     {
         extern(MSVC) private static ZonedDateTime FileTimeToZoned(uint fileTimeHigh, uint fileTimeLow)
